requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
docopt==0.6.2
filelock==3.12.0
googlemaps==4.10.0
idna==3.4
pipenv==2023.4.20
pipreqs==0.4.13
platformdirs==3.3.0
python-dotenv==1.0.0
requests==2.28.2
urllib3==1.26.15
virtualenv==20.22.0
virtualenv-clone==0.5.7
yarg==0.1.9
